+++ Resources/Constants.py
@@ -46,6 +46,8 @@ class Constants:
         self.os_support = 11.0
         self.min_os_support = 11.0
         self.max_os_support = 11.0
+        self.metal_build = False
+        self.wifi_build = False
 
     # Payload Location
     # OpenCore