+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.0.15
 - Add user-configurable OpenCore DEBUG builds
+- Add user-configurable Wifi and GPU patches
 
 ## 0.0.14
 - Enable ThirdPartyDrives to aid with hibernation on 3rd party SATA drives